requirements.txt
coloredlogs==14.0
humanfriendly==8.2
pyreadline==2.1
pyserial==3.4
PySide2==5.15.0
QtPy==1.9.0
shiboken2==5.15.0